edit

+++ lib/features/add_fatoura/data/datasources/addFatoura_dataSource.dart
@@ -23,7 +23,7 @@ abstract class AddfatouraDatasource {
     required int daman,
     required String damanDate,
     required String notes,
-    required int price,
+    required double price,
     required int reminder,
     required XFile image,
   });
@@ -48,7 +48,7 @@ class AddFatouraRemoteDataSource extends AddfatouraDatasource {
     required int daman,
     required String damanDate,
     required String notes,
-    required int price,
+    required double price,
     required int reminder,
     required XFile image,
   }) async {
